CountDown component created for reusability added to newItems
